% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylatex.R
\name{acs}
\alias{acs}
\title{Demographic data of 857 patients with ACS}
\format{A data frame with 857 rows and 17 variables:
\describe{
 \item{age}{patient age in years}
 \item{sex}{"Male" or "Female"}
 \item{cardiogenicShock}{"No" or "Yes"}
 \item{entry}{vascular access route, either "Femoral" or "Radial"}
 \item{Dx}{Final diagnosis, One of the followings : STEMI, NSTEMI or Unstable Angina}
 \item{EF}{ejection fraction, percentage by echocardiography}
 \item{height}{height in centimeter}
 \item{weight}{weight in kilogram}
 \item{BMI}{body mass index in kg/m2}
 \item{obesity}{obesity, "No" or "Yes"}
 \item{TC}{total cholesterol level in mg/dL}
 \item{LDLC}{low density lipoprotein cholesterol level in mg/dL}
 \item{HDLC}{high density lipoprotein cholesterol level in mg/dL}
 \item{TG}{triglyceride level in mg/dL}
 \item{DM}{history of diabetes mellitus,"No" or "Yes"}
 \item{HBP}{history of hypertension,"No" or "Yes"}
 \item{smoking}{history of smoking, One of the followings : "Never","Ex-smoker","Smoker"}
}}
\description{
A dataset containing demographic data and laboratory data of 857 patients with
acute coronary syndrome(ACS).
}
