% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelDf.R
\name{addLabelDf}
\alias{addLabelDf}
\title{Add labels to the data.frame}
\usage{
addLabelDf(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame
}
\description{
Add labels to the data.frame
}
