% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{cat.test}
\alias{cat.test}
\title{Perform chisq.test or fisher test}
\usage{
cat.test(x, ...)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{...}{Further arguments to be paseed to chisq.test or fisher.test}
}
\description{
Perform chisq.test or fisher test
}
