% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{cbind.mytable}
\alias{cbind.mytable}
\title{cbind function for class "mytable"}
\usage{
cbind.mytable(..., caption, y = NULL)

cbind.mytable(..., caption, y = NULL)
}
\arguments{
\item{...}{Objects of class "mytable", a result of a call to \code{\link{mytable}}}

\item{caption}{Unique values of grouping vaiabled used for column name of table}

\item{y}{Names of grouping variables used for caption of table}

\item{...}{Objects of class "mytable", a result of a call to \code{\link{mytable}}}

\item{caption}{Unique values of grouping vaiabled used for column name of table}

\item{y}{Names of grouping variables used for caption of table}
}
\description{
cbind function for class "mytable"

cbind function for class "mytable"
}

