% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mytable.R
\name{centerprint}
\alias{centerprint}
\title{Internal mytable functions}
\usage{
centerprint(x, ..., width = 10)
}
\arguments{
\item{x}{a character vector}

\item{...}{further arguments passed to or from other methods.}

\item{width}{an interger}
}
\description{
Internal mytable functions
These are not to be called by the user
}

