% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelDf.R
\name{changeColnameLabel}
\alias{changeColnameLabel}
\title{Add column labels to the data.frame}
\usage{
changeColnameLabel(data, formula = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{formula}{A formula}
}
\value{
A data.frame
}
\description{
Add column labels to the data.frame
}
