% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComma.R
\name{comma}
\alias{comma}
\title{Convert number to formatted number}
\usage{
comma(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Further arguments to be passed to function format}
}
\description{
Convert number to formatted number
}
