% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{completeTable}
\alias{completeTable}
\title{Make complete table}
\usage{
completeTable(y, x, data, origData)
}
\arguments{
\item{y}{character Another name of column}

\item{x}{character A name of column}

\item{data}{A data.frame with some limited data}

\item{origData}{A data.frame with full data}
}
\value{
a table
}
\description{
Make complete table
}
