% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressMytable.R
\name{deleteRows}
\alias{deleteRows}
\title{Delete rows of class mytable object}
\usage{
deleteRows(x, rows)
}
\arguments{
\item{x}{An object of class mytable or cbind.mytable}

\item{rows}{rows to delete}
}
\description{
Delete rows of class mytable object
}
