% Generated by roxygen2 (4.0.2): do not edit by hand
\name{densityplot}
\alias{densityplot}
\title{Make Kernel density plot}
\usage{
densityplot(formula, data, main = "", xlab = "", ylab = "", ...)
}
\arguments{
\item{formula}{an R model formula, of the form ~ variable to estimate the
unconditional density of variable, or variable ~ factor to estimate the
density of variable within each level of factor.}

\item{data}{an optional data frame containing the data.}

\item{main}{main title of plot}

\item{xlab}{label for the horizontal-axis; defaults to the name of the
variable x.}

\item{ylab}{label for the vertical axis; defaults to "Density".}

\item{...}{arguments to be passed to plot}
}
\value{
This function return NULL invisibly and draw graphs.
}
\description{
Make Kernel density plot
}
\examples{
require(moonBook)
  data(acs)
  densityplot(age~Dx,data=acs)
}

