% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractHR}
\alias{extractHR}
\title{Extract hazard ratio from a data.frame}
\usage{
extractHR(x, digits = 2)
}
\arguments{
\item{x}{an object of class coxph}

\item{digits}{An integer indicating the number of decimal places (round) or
significant digits (signif) to be used. Default value is 2.}
}
\value{
a data.frame consist of hazard ratio and 95% confidence intervals and
        the p values.
}
\description{
Extract hazard ratio from a data.frame
}

