% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressMytable.R
\name{extractKind}
\alias{extractKind}
\title{Extract kind of an object of class mytable}
\usage{
extractKind(df)
}
\arguments{
\item{df}{An object of class mytable or cbind.mytable}
}
\description{
Extract kind of an object of class mytable
}
