% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractOR}
\alias{extractOR}
\title{Extract the odds ratios from a S3 object of glm}
\usage{
extractOR(x, digits = 2)
}
\arguments{
\item{x}{A S3 object of glm}

\item{digits}{An integer indicating the number of decimal places (round) or
significant digits (signif) to be used. Default value is 2.}
}
\value{
A data.frame consist of odds ratios and 95% confidence interval and
        p values
}
\description{
Extract the odds ratios from a S3 object of glm
}

