% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelDf.R
\name{extractVars}
\alias{extractVars}
\title{Extract varaiable names from formula}
\usage{
extractVars(formula, data)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}
}
\description{
Extract varaiable names from formula
}
