% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelDf.R
\name{getLabel}
\alias{getLabel}
\title{Add column labels to the data.frame}
\usage{
getLabel(data, colname, use.column.label = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{colname}{character. column name}

\item{use.column.label}{Logical. Whether or not use column labels.}
}
\description{
Add column labels to the data.frame
}
