% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{my.t.test}
\alias{my.t.test}
\title{Internal mytable functions}
\usage{
my.t.test(y, x)

my.t.test(y, x)
}
\arguments{
\item{y}{a vector}

\item{x}{a numeric vector}

\item{y}{a vector}

\item{x}{a numeric vector}
}
\description{
Internal mytable functions
These are not to be called by the user

Internal mytable functions
These are not to be called by the user
}

