% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mycsv.cbind.mytable}
\alias{mycsv.cbind.mytable}
\title{Export to csv file for class "cbind.mytable"}
\usage{
\method{mycsv}{cbind.mytable}(x, row.names = FALSE, ...)
}
\arguments{
\item{x}{An object of class "cbind.mytable" a result of a call to \code{\link{mytable}}}

\item{row.names}{either a logical value indicating whether the row names of x
are to be written along with x, or a character vector of
row names to be written.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Export to csv file for class "cbind.mytable"
}
\examples{
require(moonBook)
res1=mytable(sex+Dx~age+DM,data=acs)
mycsv(res1,file="test1.csv")
mycsv(summary(res1),file="testsummary1.csv")
}

