% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mycsv}
\alias{mycsv}
\title{Export to csv file for class "mytable" or "cbind.mytable"}
\usage{
mycsv(x, row.names = FALSE, ...)
}
\arguments{
\item{x}{An object of class "mytable" or "cbind.mytable"}

\item{row.names}{either a logical value indicating whether the row names of x
are to be written along with x, or a character vector of
row names to be written.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Export to csv file for class "mytable" or "cbind.mytable"
}
\examples{
require(moonBook)
res=mytable(sex~age+DM,data=acs)
mycsv(res,file="test.csv")
mycsv(summary(res),file="testsummary.csv")
}

