% Generated by roxygen2 (4.0.2): do not edit by hand
\name{myhtml.mytable}
\alias{myhtml.mytable}
\title{Export to html file for class "mytable" or "cbind.mytable"}
\usage{
\method{myhtml}{mytable}(x, caption = NULL)
}
\arguments{
\item{x}{An object of class "mytable" or "cbind.mytable"}

\item{caption}{A character}
}
\description{
Export to html file for class "mytable" or "cbind.mytable"
}
\examples{
require(moonBook)
res=mytable(sex~age+Dx,data=acs)
myhtml(res)
}

