% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mylatex.mytable}
\alias{mylatex.mytable}
\title{Exporting "mytable" table to LaTeX format}
\usage{
\method{mylatex}{mytable}(myobj, size = 5, caption = NULL)
}
\arguments{
\item{myobj}{An object of class 'mytable'}

\item{size}{An integer indicating font size, defaulting is 5.}

\item{caption}{A character}
}
\description{
This function takes the result of mytable and exports the tables to LaTeX format.
}
\examples{
out=mytable(am~.,data=mtcars)
mylatex(out)
}

