% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mylatex.data.frame}
\alias{mylatex}
\alias{mylatex.cbind.mytable}
\alias{mylatex.data.frame}
\alias{mylatex.default}
\alias{mylatex.mytable}
\title{Exporting "cbind.mytable","mytable","data.frame" to LaTeX format}
\usage{
\method{mylatex}{data.frame}(myobj, size = 5, caption = NULL,
  rownames = TRUE, caption.placement = "top", caption.position = "c",
  align = NULL, digits = NULL)

mylatex(myobj, size = 5, caption = NULL, rownames = TRUE,
  caption.placement = "top", caption.position = "c", align = NULL,
  digits = NULL)

\method{mylatex}{default}(myobj, size = 5, caption = NULL,
  rownames = TRUE, caption.placement = "top", caption.position = "c",
  align = NULL, digits = NULL)

\method{mylatex}{mytable}(myobj, size = 5, caption = NULL,
  rownames = TRUE, caption.placement = "top", caption.position = "c",
  align = NULL, digits = NULL)

\method{mylatex}{cbind.mytable}(myobj, size = 5, caption = NULL,
  rownames = TRUE, caption.placement = "top", caption.position = "c",
  align = NULL, digits = NULL)
}
\arguments{
\item{myobj}{An object of class 'mytable'}

\item{size}{An integer indicating font size, defaulting is 5.}

\item{caption}{A character}

\item{rownames}{A logical value wheher or not include rownames in table}

\item{caption.placement}{The caption will be have placed at the top of the table
if caption.placement is "top" and at the bottom of the table
if it equals "bottom". Default value is "top".}

\item{caption.position}{The caption will be have placed at the center of the table
if caption.position is "center" or "c", and at the left side of the table
if it equals "left" or "l", and at the right side of the table
if it equals "right" or "r". Default value is "center".}

\item{align}{Character vector : nchar equal to the number of columns of the
resulting table indicating the alignment of the corresponding columns.}

\item{digits}{Numeric vector of length equal to one (in which case it will be
replicated as necessary) or to the number of columns of the resulting table}
}
\description{
Exporting "cbind.mytable","mytable","data.frame" to LaTeX format
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{default}: 

\item \code{mytable}: 

\item \code{cbind.mytable}: 
}}
\examples{
require(moonBook)
out=mytable(sex~.,data=acs)
mylatex(out)
out1=mytable(sex+Dx~.,data=acs)
mylatex(out1,size=6)
x=head(iris)
mylatex(x,size=3,caption="Table 1. mylatex Test")
mylatex(x,size=7,caption="Table 1. mylatex Test",caption.position="l")
mylatex(x,size=7,caption="Table 1. mylatex Test",caption.placement="bottom",
      caption.position="l")
}

