% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{mytable.sub}
\alias{mytable.sub}
\title{Internal mytable functions}
\usage{
mytable.sub(y, x, data, max.ylev = 5, method = 1, show.total = FALSE)

mytable.sub(y, x, data, max.ylev = 5, method = 1, show.total = FALSE)
}
\arguments{
\item{y}{a vector}

\item{x}{a vector}

\item{data}{a data.frame}

\item{max.ylev}{an integer}

\item{method}{an integer}

\item{show.total}{a logical value}

\item{y}{a vector}

\item{x}{a vector}

\item{data}{a data.frame}

\item{max.ylev}{an integer}

\item{method}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user

Internal mytable functions
These are not to be called by the user
}

