% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mytable2df}
\alias{mytable2df}
\title{Convert mytable object to data.frame}
\usage{
mytable2df(x)
}
\arguments{
\item{x}{An object of class "mytable" a result of a call to \code{\link{mytable}}}
}
\value{
a data.frame with N number
}
\description{
Add N number into data.frame
}

