% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhtml.R
\name{mytable2html}
\alias{mytable2html}
\title{Prepare mytable object to data.frame ready to html}
\usage{
mytable2html(x)
}
\arguments{
\item{x}{An object of class "mytable" a result of a call to \code{\link{mytable}}}
}
\value{
a data.frame with N number
}
\description{
Add N number into data.frame
}
