% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.df.R
\name{mytable_df}
\alias{mytable_df}
\title{make mytable from data.frame}
\usage{
mytable_df(x, max.ylev = 5, maxCatLevel = 20, digits = 1, method = 1,
  show.all = FALSE)
}
\arguments{
\item{x}{A data.frame}

\item{max.ylev}{An integer indicating the maximum number of levels of grouping
variable ('y'). If a colummn have unique values less than max.ylev
it is treated as a categorical variable. Default value is 5.}

\item{digits}{An integer indicating the number of decimal places (round) or
significant digits to be used. Default value is 1.}

\item{method}{An integer indicating methods for continuous variables.
Possible values in methods are
\describe{
   \item{1}{forces analysis as normal-distributed}
   \item{2}{forces analysis as continuous non-normal}
   \item{3}{performs a Shapiro-Wilk test to decide between
           normal or non-normal}
}
Default value is 1.}

\item{show.all}{A logical value indicating whether or not all statistical
values have to be shown in table. Default value is FALSE.}
}
\value{
An object of class "mytable.df".
     'print' returns a table for descriptive statistics.
}
\description{
make mytable from data.frame
}
