% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{mytable_sub2}
\alias{mytable_sub2}
\title{Internal mytable functions}
\usage{
mytable_sub2(
  y,
  x,
  data,
  max.ylev = 5,
  maxCatLevel = 20,
  method = 1,
  catMethod = 2,
  show.total = FALSE,
  origData
)
}
\arguments{
\item{y}{a vector}

\item{x}{a vector}

\item{data}{a data.frame}

\item{max.ylev}{an integer}

\item{maxCatLevel}{an integer}

\item{method}{an integer}

\item{catMethod}{an integer}

\item{show.total}{a logical value}

\item{origData}{a data.frame}
}
\description{
Internal mytable functions
These are not to be called by the user
}
