% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{num_summary}
\alias{num_summary}
\title{Internal mytable functions}
\usage{
num_summary(x)

num_summary(x)
}
\arguments{
\item{x}{a numeric vector}

\item{x}{a numeric vector}
}
\description{
Internal mytable functions
These are not to be called by the user

Internal mytable functions
These are not to be called by the user
}

