% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mytable.R
\name{obj2linecount}
\alias{obj2linecount}
\title{Internal mytable functions}
\usage{
obj2linecount(myobj)
}
\arguments{
\item{myobj}{an R object}
}
\description{
Internal mytable functions
These are not to be called by the user
}

