% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.df.R
\name{print.mytable.df}
\alias{print.mytable.df}
\title{Print an object of mytable.df}
\usage{
\method{print}{mytable.df}(x, ...)
}
\arguments{
\item{x}{An object of class mytable.df}

\item{...}{Further arguments}
}
\description{
Print an object of mytable.df
}
