% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{print.mytable}
\alias{print.mytable}
\title{Print function for class "mytable"}
\usage{
\method{print}{mytable}(x, ...)

\method{print}{mytable}(x, ...)
}
\arguments{
\item{x}{An object of class "mytable", a result of a call to \code{\link{mytable}}}

\item{...}{further arguments passed to or from other methods.}

\item{x}{An object of class "mytable", a result of a call to \code{\link{mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for class "mytable"

Print function for class "mytable"
}

