% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{printmytable2}
\alias{printmytable2}
\title{Internal mytable functions}
\usage{
printmytable2(obj, digits = 1, useSE = FALSE)
}
\arguments{
\item{obj}{an object}

\item{digits}{an integer}

\item{useSE}{logical}
}
\description{
Internal mytable functions
These are not to be called by the user
}
