% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mylatex.R
\name{radial}
\alias{radial}
\title{Demographic data of 115 patients performing IVUS(intravascular ultrasound)
examination of a radial artery.}
\format{A data frame with 115 rows and 15 variables:
\describe{
 \item{male}{if Male, 1; if Female 0}
 \item{age}{patient age in years}
 \item{height}{height in centimeter}
 \item{weight}{weight in kilogram}
 \item{HBP}{history of hypertension, 1 for yes or 0 for no}
 \item{DM}{history of diabetes mellitus, 1 for yes or 0 for no}
 \item{smoking}{history of smoking, One of the followings : "non-smoker","ex-smoker","smoker"}
 \item{TC}{total cholesterol level in mg/dL}
 \item{TG}{triglyceride level in mg/dL}
 \item{HDL}{high density lipoprotein cholesterol level in mg/dL}
 \item{LDL}{low density lipoprotein cholesterol level in mg/dL}
 \item{hsCRP}{high-sensitive C reactive protein}
 \item{NTAV}{normalized total atheroma volume measured by IVUS in cubic mm}
 \item{PAV}{percent atheroma volume in percentage}
 \item{sex}{Factor with two levels; "Male" or "Female"}
}}
\description{
A dataset containing demographic data and laboratory data of 115 pateints performing IVUS(intravascular ultrasound)
examination of a radial artery after tansradial coronary angiography.
}

