% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{rank2group}
\alias{rank2group}
\title{rank a numeric vector and returns a new ordinal vector}
\usage{
rank2group(y, k = 4)
}
\arguments{
\item{y}{a numeric vector}

\item{k}{a integer specifies how many groups you want to classify.
default value is 4}
}
\value{
a ordinal vector(numeric) with the same length of y
}
\description{
rank a numeric vector and returns a new ordinal vector
}
\examples{

require(ggplot2)
data(diamonds)
diamonds$PriceGroup=rank2group(diamonds$price,4)
table(diamonds$PriceGroup)
aggregate(price~PriceGroup,data=diamonds,range)

diamonds$PriceGroup3=rank2group(diamonds$price,3)
table(diamonds$PriceGroup3)
aggregate(price~PriceGroup3,data=diamonds,range)
diamonds$PriceGroup5=rank2group(diamonds$price,5)
table(diamonds$PriceGroup5)
aggregate(price~PriceGroup5,data=diamonds,range)
}
