% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{reprint}
\alias{reprint}
\title{Internal mytable functions}
\usage{
reprint(x, times)

reprint(x, times)
}
\arguments{
\item{x}{a chearacter vector}

\item{times}{an integer}

\item{x}{a chearacter vector}

\item{times}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user

Internal mytable functions
These are not to be called by the user
}

