% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmytable.R, R/mytable.R
\name{space}
\alias{space}
\title{Internal mytable functions}
\usage{
space(num)

space(num)
}
\arguments{
\item{num}{an integer}

\item{num}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user

Internal mytable functions
These are not to be called by the user
}

