% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{std.error}
\alias{std.error}
\title{calculate stendard error}
\usage{
std.error(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical}
}
\description{
calculate stendard error
}
\examples{
x=c(179,160,136,227,123,23,45,67,1,234)
std.error(x)
}
