% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mytable.R
\name{summary.cbind.mytable}
\alias{summary.cbind.mytable}
\title{Summarizing function for class "cbind.mytable"}
\usage{
\method{summary}{cbind.mytable}(object, ...)
}
\arguments{
\item{object}{An object of class "cbind.mytable", a result of a call \code{\link{mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarizing function for class "cbind.mytable"
}
\examples{
out=mytable(am+cyl~.,data=mtcars)
summary(out)
}

