% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.mytable}
\alias{summary.mytable}
\title{Summarizing function for class "mytable"}
\usage{
\method{summary}{mytable}(object, ...)
}
\arguments{
\item{object}{An object of class "mytable", a result of a call \code{\link{mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarizing function for class "mytable"
}
\examples{
out=mytable(am~.,data=mtcars)
summary(out)
}

