% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validColname}
\alias{validColname}
\title{Find valid string among character vecter from approximate string}
\usage{
validColname(pattern, x)
}
\arguments{
\item{pattern}{character string to be matched in the given character}

\item{x}{a character vecter where matches are sought}
}
\value{
returns NA in case of no matched string found
        or a character string in string vecter x
}
\description{
Find valid string among character vecter from approximate string
}
\examples{
a="dx"
b=c("Age","Sex","Dx")
validColname(a,b)
}

