% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComma.R
\name{addComma}
\alias{addComma}
\alias{addComma.mytable}
\alias{addComma.mytable.df}
\alias{addComma.cbind.mytable}
\alias{addComma.data.frame}
\alias{addComma.character}
\title{Change numbers into formatted numbers}
\usage{
addComma(x)

\method{addComma}{mytable}(x)

\method{addComma}{mytable.df}(x)

\method{addComma}{cbind.mytable}(x)

\method{addComma}{data.frame}(x)

\method{addComma}{character}(x)
}
\arguments{
\item{x}{An object}
}
\description{
Change numbers into formatted numbers
}
\section{Methods (by class)}{
\itemize{
\item \code{mytable}: S3 method for class mytable

\item \code{mytable.df}: S3 method for class mytable.df

\item \code{cbind.mytable}: S3 method for class cbind.mytable

\item \code{data.frame}: S3 method for class data.frame

\item \code{character}: S3 method for class character
}}

\examples{
\dontrun{
require(stringr)
require(magrittr)
require(ggplot2)
mytable(cut~.,data=diamonds) \%>\% addComma
x=mytable(Dx~sex,data=acs)
addComma(x)
}
}
