% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbind.mytable}
\alias{cbind.mytable}
\title{cbind function for class "mytable"}
\usage{
cbind.mytable(..., caption, y = NULL)
}
\arguments{
\item{...}{Objects of class "mytable", a result of a call to \code{\link{mytable}}}

\item{caption}{Unique values of grouping vaiabled used for column name of table}

\item{y}{Names of grouping variables used for caption of table}
}
\description{
cbind function for class "mytable"
}

