% Generated by roxygen2 (4.0.2): do not edit by hand
\name{centerprint}
\alias{centerprint}
\title{Internal mytable functions}
\usage{
centerprint(x, ..., width = 10)
}
\arguments{
\item{x}{a character vector}

\item{...}{further arguments passed to or from other methods.}

\item{width}{an interger}
}
\description{
Internal mytable functions
These are not to be called by the user
}

