% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelDf.R
\name{changeColnameLabel}
\alias{changeColnameLabel}
\title{Change column names with labels}
\usage{
changeColnameLabel(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Change column names with labels
}
