% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my.chisq.test}
\alias{my.chisq.test}
\title{Internal mytable functions}
\usage{
my.chisq.test(x, y, mydata)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{mydata}{a data.frame}
}
\description{
Internal mytable functions
These are not to be called by the user
}

