% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhtml.R
\name{myhtml}
\alias{myhtml}
\alias{myhtml.default}
\alias{myhtml.mytable}
\alias{myhtml.cbind.mytable}
\title{Export to html file for class "mytable" or "cbind.mytable" of "data.frame"}
\usage{
myhtml(x, caption = NULL, rownames = TRUE)

\method{myhtml}{default}(x, caption = NULL, rownames = TRUE)

\method{myhtml}{mytable}(x, caption = NULL, rownames = TRUE)

\method{myhtml}{cbind.mytable}(x, caption = NULL, rownames = TRUE)
}
\arguments{
\item{x}{An object of class "mytable" or "cbind.mytable"}

\item{caption}{A character}

\item{rownames}{A logical value whether or not include rownames in table}
}
\description{
Export to html file for class "mytable" or "cbind.mytable" of "data.frame"
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{mytable}: 

\item \code{cbind.mytable}: 
}}

\examples{
require(moonBook)
res=mytable(sex~age+Dx,data=acs)
myhtml(res)
res1=mytable(sex+Dx~.,data=acs)
myhtml(res1)
x=head(iris)
myhtml(x)
myhtml(x,caption="Table 1. myhtml Test")
myhtml(x,caption="Table 1. myhtml Test",rownames=FALSE)
}
