% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mylatex}
\alias{mylatex}
\title{Exporting "cbind.mytable","mytable" table to LaTeX format}
\usage{
mylatex(myobj, size = 5, caption = NULL)
}
\arguments{
\item{myobj}{An Object of class ""cbind.mytable" or "mytable"}

\item{size}{An integer indicating font size, defaulting is 5.}

\item{caption}{A character}
}
\description{
Exporting "cbind.mytable","mytable" table to LaTeX format
}
\examples{
out=mytable(am~.,data=mtcars)
mylatex(out)
out1=mytable(am+cyl~.,data=mtcars)
mylatex(out1)
}

