% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{mytable.sub}
\alias{mytable.sub}
\title{Internal mytable functions}
\usage{
mytable.sub(y, x, data, max.ylev = 5, method = 1, show.total = FALSE)
}
\arguments{
\item{y}{a vector}

\item{x}{a vector}

\item{data}{a data.frame}

\item{max.ylev}{an integer}

\item{method}{an integer}

\item{show.total}{a logical value}
}
\description{
Internal mytable functions
These are not to be called by the user
}
