% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{num_summary}
\alias{num_summary}
\title{Internal mytable functions}
\usage{
num_summary(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Internal mytable functions
These are not to be called by the user
}
