% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{p2sig}
\alias{p2sig}
\title{Internal mytable functions}
\usage{
p2sig(value)
}
\arguments{
\item{value}{a numeric vector}
}
\description{
Internal mytable functions
These are not to be called by the user
}
