% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.cbind.mytable}
\alias{print.cbind.mytable}
\title{Print function for class "cbind.mytable"}
\usage{
\method{print}{cbind.mytable}(x, ...)
}
\arguments{
\item{x}{an object of class "cbind.mytable", a result of a call to \code{\link{cbind.mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for class "cbind.mytable"
}

