% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{print.cbind.mytable}
\alias{print.cbind.mytable}
\title{Print function for class "cbind.mytable"}
\usage{
\method{print}{cbind.mytable}(x, ...)
}
\arguments{
\item{x}{an object of class "cbind.mytable", a result of a call to \code{\link{cbind.mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for class "cbind.mytable"
}
