% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{print.mytable}
\alias{print.mytable}
\title{Print function for class "mytable"}
\usage{
\method{print}{mytable}(x, ...)
}
\arguments{
\item{x}{An object of class "mytable", a result of a call to \code{\link{mytable}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for class "mytable"
}
