% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mytable.R
\name{printmytable2}
\alias{printmytable2}
\title{Internal mytable functions}
\usage{
printmytable2(obj, digits = 1)
}
\arguments{
\item{obj}{an object}

\item{digits}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user
}

