% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reprint}
\alias{reprint}
\title{Internal mytable functions}
\usage{
reprint(x, times)
}
\arguments{
\item{x}{a chearacter vector}

\item{times}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user
}

