% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{reprint}
\alias{reprint}
\title{Internal mytable functions}
\usage{
reprint(x, times)
}
\arguments{
\item{x}{a character vector}

\item{times}{an integer}
}
\description{
Internal mytable functions
These are not to be called by the user
}
