% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{validColname}
\alias{validColname}
\title{Find valid string among character vector from approximate string}
\usage{
validColname(pattern, x)
}
\arguments{
\item{pattern}{character string to be matched in the given character}

\item{x}{a character vector where matches are sought}
}
\value{
returns NA in case of no matched string found
        or a character string in string vector x
}
\description{
Find valid string among character vector from approximate string
}
\examples{
a="dx"
b=c("Age","Sex","Dx")
validColname(a,b)

}
